[REFACTOR] rename 'rpc' module to network
it will contain the legacy files as well
