+++ inexor/client/CMakeLists.txt
@@ -66,8 +66,7 @@ require_threads(${CLIENT_BINARY})
 require_crashreporter(${CLIENT_BINARY})
 require_sdl(${CLIENT_BINARY})
 require_zlib(${CLIENT_BINARY})
-require_enet(${CLIENT_BINARY})
-require_rpc(${CLIENT_BINARY} "CLIENT NOT_STANDALONE")
+require_network(${CLIENT_BINARY} "CLIENT NOT_STANDALONE")
 require_util(${CLIENT_BINARY})
 require_ui(${CLIENT_BINARY})
 require_texture(${CLIENT_BINARY})
