+++ inexor/network/SharedOptions.hpp
