+++ inexor/network/SharedTree.hpp
@@ -1,0 +1,7 @@
+#pragma once
+
+#include "inexor/network/SharedOptions.hpp"
+#include "inexor/network/SharedVar.hpp"
+#include "inexor/network/SharedClass.hpp"
+#include "inexor/network/SharedFunc.hpp"
+
