--- inexor/rpc/SharedTree.hpp
@@ -1,7 +1,0 @@
-#pragma once
-
-#include "inexor/rpc/SharedOptions.hpp"
-#include "inexor/rpc/SharedVar.hpp"
-#include "inexor/rpc/SharedClass.hpp"
-#include "inexor/rpc/SharedFunc.hpp"
-
