+++ inexor/server/CMakeLists.txt
@@ -19,8 +19,7 @@ add_app(${SERVER_BINARY} ${SERVER_SOURCES})
 
 require_threads(${SERVER_BINARY})
 require_zlib(${SERVER_BINARY})
-require_enet(${SERVER_BINARY})
-require_rpc(${SERVER_BINARY} "SERVER STANDALONE SERVMODE")
+require_network(${SERVER_BINARY} "SERVER STANDALONE SERVMODE")
 require_util(${SERVER_BINARY})
 require_crashreporter(${SERVER_BINARY})
 require_filesystem(${SERVER_BINARY})
