+++ inexor/ui/screen/ScreenManager.hpp
@@ -2,7 +2,7 @@
 
 #include <SDL.h>
 
-#include "inexor/rpc/SharedTree.hpp"
+#include "inexor/network/SharedTree.hpp"
 
 /// Hard coded defaults for the SDL window
 #define SCR_MINW 320
